+++ src/Game.cpp
@@ -116,7 +116,7 @@ int Game::loadLevel(const char* level) {
             Render::setMode(RenderMode::Texture);
 
             Camera::setPosition(glm::vec3(getLara().getPos(0),
-                                          getLara().getPos(1) + 1024.0f,
+                                          getLara().getPos(1) - 1024.0f,
                                           getLara().getPos(2)));
         }
     } else {