+++ src/RoomMesh.cpp
@@ -72,9 +72,7 @@ void RoomMesh::prepare() {
     textures = std::move(tex);
 }
 
-void RoomMesh::display(glm::mat4 model, glm::mat4 view, glm::mat4 projection) {
-    glm::mat4 MVP = projection * view * model;
-
+void RoomMesh::display(glm::mat4 MVP) {
     if (indices.size() > 0) {
         unsigned int indexStart = 0;
         unsigned int indexPos = 1;